backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-api-python-client==2.100.0
google-analytics-data==0.18.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
requests==2.31.0
cachetools==5.3.1
